Mostly tidying up
I've tidied up the majority of the code, removing the need for the SetUp and TearDown functions.

I've commented out the method that's being used to change users terminal colors. In practise, this should work, but I think with one shell environment being different to the next, that this may not be completely portable.